refactor(server): split bootstrap resident wait bridge
